LPS-155334 Merge FDS top components and contexts
