+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/management_bar/components/ActiveViewSelector.js
@@ -17,16 +17,14 @@ import ClayDropDown from '@clayui/drop-down';
 import ClayIcon from '@clayui/icon';
 import React, {useContext, useState} from 'react';
 
-import {AppContext} from '../../AppContext';
-import DataSetContext from '../../DataSetContext';
+import FrontendDataSetContext from '../../FrontendDataSetContext';
 import persistActiveView from '../../thunks/persistActiveView';
 import ViewsContext from '../../views/ViewsContext';
 
 function ActiveViewSelector({views}) {
-	const {appURL, portletId} = useContext(AppContext);
 	const [active, setActive] = useState(false);
 	const [{activeView}, dispatch] = useContext(ViewsContext);
-	const {id} = useContext(DataSetContext);
+	const {appURL, id, portletId} = useContext(FrontendDataSetContext);
 
 	return (
 		<ClayDropDown
