+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/views/cards/Cards.js
@@ -21,7 +21,7 @@ import React, {useContext} from 'react';
 
 import {handleAction, isLink} from '../../actions/Actions';
 
-function Cards({dataLoading, dataSetContext, items, schema}) {
+function Cards({dataLoading, frontendDataSetContext, items, schema}) {
 	const {
 		executeAsyncItemAction,
 		highlightItems,
@@ -32,7 +32,7 @@ function Cards({dataLoading, dataSetContext, items, schema}) {
 		selectedItemsKey,
 		selectedItemsValue,
 		style,
-	} = useContext(dataSetContext);
+	} = useContext(frontendDataSetContext);
 
 	if (dataLoading) {
 		return <ClayLoadingIndicator className="mt-7" />;
